feat(monitor): ship plugin monitor via npx subcommand; reconcile docs
The plugin's experimental.monitors[] entry ran
`node ${CLAUDE_PLUGIN_ROOT}/dist/monitor/index.js`, but dist/ is
gitignored — a github-marketplace plugin clone carries no built monitor
binary, so the monitor never started via the persistent-install path.
Ship it through `npx -y tandem-editor@<v> monitor` (npm ships dist via
files:["dist/"]), matching the tandem / tandem-channel mcpServers entries.

Split src/monitor/index.ts into a thin standalone entry + a shared runtime
(run.ts) so the `tandem monitor` subcommand can import the runtime WITHOUT
the isDirectRun auto-run: in the bundled dist/cli/index.js that guard
resolves TRUE and would fire main() a second time (doubled SSE
subscription emitting every event twice). The import is dynamic and
in-branch so the runtime's console->stderr redirect never runs for other
commands.

Drop the redundant env block from the monitor entry: resolveTandemUrl()
already defaults to the same loopback URL, and ADR-028 records that the
monitors[] manifest schema rejected env blocks — keeping it risks
regressing the activation the 2.1.212 re-test proved.

Correct now-false "monitor cannot activate via any path" claims across the
architecture/decisions/roadmap/setup/apply docs and the two spike reports:
the 2.1.143 NO-GO was reversed on Claude Code 2.1.212.

Note: the npx pin is release-coupled. Published 0.17.0 has no `monitor`
subcommand, so the manifest entry only becomes functional once a release
that publishes it lands; the six-surface version bump advances package.json
and the pin together.

Tests: structural guard for the entry/runtime split + a spawn test
asserting `tandem monitor` starts main() exactly once; extend the
version-pin guard to walk experimental.monitors[] npx commands.
