+++ .claude-plugin/plugin.json
@@ -28,7 +28,7 @@
     "monitors": [
       {
         "name": "tandem-events",
-        "command": "node ${CLAUDE_PLUGIN_ROOT}/dist/monitor/index.js",
+        "command": "npx -y tandem-editor@0.17.0 monitor",
         "description": "Tandem real-time document events (annotations, chat, selections)"
       }
     ]
