+++ docs/architecture.md
@@ -349,7 +349,7 @@ When Claude Code asks for tool approval, it sends `notifications/claude/channel/
 
 ## Plugin Monitor
 
-> **Forward-looking, not the live path (Spike B / #985).** The plugin monitor is the *intended* future replacement for the channel shim, but Claude Code does not activate `experimental.monitors[].command` via any path Tandem can use today (`--plugin-dir`, path-source install, or github-marketplace install — see `docs/spikes/plugin-monitor-viability-spike.md`). Until Claude Code surfaces a working activation path, **the channel shim is the actual v1.0 Claude Code push transport**, and `tandem setup` registers it by default for the Claude Code target. The design below describes the monitor's role once activation lands.
+> **Activates on Claude Code 2.1.212+; the channel shim is still the default (Spike B / #985 NO-GO, since reversed).** The plugin monitor was found inactive on Claude Code 2.1.143 (the historical Spike B NO-GO — `docs/spikes/plugin-monitor-viability-spike.md`, whose probes also ran in `-p` print mode, where monitors never activate by design). It **does** activate on 2.1.212+ interactive sessions via `--plugin-dir` and persistent installs (re-tested 2026-07-17), and it ships installable via `npx -y tandem-editor@<version> monitor` (the manifest previously ran `node ${CLAUDE_PLUGIN_ROOT}/dist/monitor/index.js`, but `dist/` is gitignored so a github-marketplace clone carried no monitor binary; npm ships `dist`, so npx delivers it). It is an independent push path needing no `--dangerously-...` flag. Which of the monitor and the channel shim becomes the *canonical* transport is an open decision — both active in one session double-deliver — so `tandem setup` still registers the channel shim by default for the Claude Code target. The design below describes the monitor's role.
 
 The plugin monitor (`src/monitor/index.ts`) is the planned modern alternative to the channel shim for receiving real-time events from Tandem. It is installed as a Claude Code plugin rather than spawned as a stdio subprocess.
