+++ docs/decisions.md
@@ -331,6 +331,8 @@ Probe instrumentation — `src/server/mcp/server.ts` patched to (a) advertise `b
 **Consequences:** Existing installs are backward compatible — all env vars are optional; callers get the same defaults when absent. Cowork real-time push remains gated on Sub-task D verification. `monitors[]` manifest `env` support remains an upstream gap — file issue against `anthropics/claude-code` (ref: issue #52245 monitor auto-arm, issue #27398 Cowork hook gaps).
 **Update (2026-06-02, #985):** Spike B (`docs/spikes/plugin-monitor-viability-spike.md`, #712) subsequently established that Claude Code does not activate `experimental.monitors[].command` via *any* install path Tandem can use (`--plugin-dir`, path-source, or github-marketplace). The plugin monitor is therefore forward-looking only; until upstream activation lands, the **channel shim is the canonical v1.0 Claude Code push transport** and is registered by default for the Claude Code target (`shouldRegisterChannelShim`). This reverts the earlier de-facto "monitor-canonical, channel-off-by-default" posture — see the #985 CHANGELOG entry and `docs/architecture.md` Plugin Monitor section.
 
+**Update (2026-07-17):** Re-tested on Claude Code **2.1.212** — the monitor **does** activate interactively (via `--plugin-dir` and persistent installs), delivering an event to an idle session with no `--dangerously-...` flag. The 2.1.143 NO-GO above was version-specific, and its probes ran in `-p` print mode, where monitors never activate by design — so the NO-GO conflated two confounds. The monitor now ships **installable** via `npx -y tandem-editor@<version> monitor`: the manifest previously ran `node ${CLAUDE_PLUGIN_ROOT}/dist/monitor/index.js`, but `dist/` is gitignored so a github clone carried no binary, while npm ships `dist`. The channel shim **remains the registered default**; the monitor is an independent push path, and the canonical-transport choice (channel vs monitor) is **deferred** pending a content-richness probe (does the plugin host surface the monitor's stdout body or a generic wake?) and a measured double-fire rate (both active in one session double-deliver, confirmed).
+
 ## ADR-029: Action Registry and Command Palette
 
 **Status:** Accepted
