+++ docs/spikes/marketplace-install-spike.md
@@ -1,6 +1,8 @@
 # Marketplace Install Spike (F2 — ADR-038 Phase 3 gate)
 
-**Status:** Spike complete — **NO-GO on marketplace one-command monitor activation in v2.1.143.** Re-validated 2026-05-18 against current master (`.claude-plugin/plugin.json` + `marketplace.json` unchanged since the spike; Claude Code still v2.1.143). Verdict holds.
+> **⚠️ SUPERSEDED (2026-07-17).** This NO-GO was doubly confounded: it ran in `-p` **print mode** (where monitors never activate by design) on 2.1.143. Two later findings moot it: (1) the monitor **does** activate on Claude Code 2.1.212 interactive sessions (see the reversal banner in `plugin-monitor-viability-spike.md`); (2) the marketplace path also failed for a **packaging** reason — the manifest ran `node ${CLAUDE_PLUGIN_ROOT}/dist/monitor/index.js`, but `dist/` is gitignored, so a github-marketplace clone carried no monitor binary regardless of activation. Fixed by shipping the monitor via `npx -y tandem-editor@<version> monitor` (npm ships `dist`). The body below is preserved as the historical record.
+
+**Status:** Spike complete — **NO-GO on marketplace one-command monitor activation in v2.1.143.** Re-validated 2026-05-18 against current master (`.claude-plugin/plugin.json` + `marketplace.json` unchanged since the spike; Claude Code still v2.1.143). Verdict holds. *(superseded 2026-07-17 — see banner above)*
 **Date:** 2026-05-17 (initial); 2026-05-18 (re-validation)
 **Claude Code version tested:** 2.1.143
 **Refs:** [ADR-038](../decisions.md#adr-038-mcp-first-integration-policy-claude-as-default-integration), [Plugin Monitor Viability Spike (Spike B)](./plugin-monitor-viability-spike.md), [#477](https://github.com/bloknayrb/tandem/issues/477).
